Вынес FeedbackFormQueryParams и TaskInfoPageQueryParams в отдельный файл.
