+++ src/api/schemas/__init__.py
@@ -7,11 +7,9 @@ from .feedback import FeedbackSchema
 from .health_check import BotStatus, CommitStatus, DBStatus, HealthCheck
 from .notification import (
     ErrorsSending,
-    FeedbackFormQueryParams,
     InfoRate,
     Message,
     MessageList,
-    TaskInfoPageQueryParams,
     TelegramNotificationRequest,
     TelegramNotificationUsersGroups,
     TelegramNotificationUsersRequest,
@@ -35,8 +33,6 @@ __all__ = (
     "DBStatus",
     "HealthCheck",
     "ErrorsSending",
-    "FeedbackFormQueryParams",
-    "TaskInfoPageQueryParams",
     "InfoRate",
     "Message",
     "MessageList",
