+++ src/bot/schemas/__init__.py
@@ -1,0 +1,6 @@
+from .web_app_query_params import FeedbackFormQueryParams, TaskInfoPageQueryParams
+
+__all__ = (
+    "FeedbackFormQueryParams",
+    "TaskInfoPageQueryParams",
+)
